# apps/server/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PracticeSessionOut": {
        "additionalProperties": false,
        "properties": {
          "anchorMeasure": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchormeasure"
          },
          "anchorPass": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchorpass"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "endedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endedat"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "leaderId": {
            "title": "Leaderid",
            "type": "string"
          },
          "repertoireId": {
            "title": "Repertoireid",
            "type": "string"
          },
          "roomId": {
            "title": "Roomid",
            "type": "string"
          },
          "serverStartTimeNs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serverstarttimens"
          },
          "status": {
            "enum": [
              "idle",
              "armed",
              "playing",
              "stopped"
            ],
            "title": "Status",
            "type": "string"
          },
          "tempoMapRevision": {
            "title": "Tempomaprevision",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "roomId",
          "repertoireId",
          "leaderId",
          "tempoMapRevision",
          "status",
          "anchorMeasure",
          "anchorPass",
          "serverStartTimeNs",
          "createdAt",
          "updatedAt",
          "endedAt"
        ],
        "title": "PracticeSessionOut",
        "type": "object"
      },
...
      "RoomCreate": {
        "additionalProperties": false,
        "properties": {
          "repertoireId": {
            "title": "Repertoireid",
            "type": "string"
          }
        },
        "required": [
          "repertoireId"
        ],
        "title": "RoomCreate",
        "type": "object"
      },
      "RoomOut": {
        "additionalProperties": false,
        "properties": {
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "leaderId": {
            "title": "Leaderid",
            "type": "string"
          },
          "repertoireId": {
            "title": "Repertoireid",
            "type": "string"
          },
          "roomId": {
            "title": "Roomid",
            "type": "string"
          },
          "tempoMapRevision": {
            "title": "Tempomaprevision",
            "type": "integer"
          }
        },
        "required": [
          "roomId",
          "repertoireId",
          "leaderId",
          "tempoMapRevision",
          "expiresAt"
        ],
        "title": "RoomOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API and WebSocket clock/transport gateway. Clients expand tempo-map beats locally; the server only agrees on revision, anchor, and absolute server time.",
    "title": "FeelMyRythm API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/practice-sessions/{session_id}": {
      "get": {
        "operationId": "get_practice_session_api_practice_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Practice Session",
        "tags": [
          "practice rooms"
        ]
      }
    },
...
    "/api/repertoire/{repertoire_id}/practice-sessions": {
      "get": {
        "operationId": "list_practice_sessions_api_repertoire__repertoire_id__practice_sessions_get",
        "parameters": [
          {
            "in": "path",
            "name": "repertoire_id",
            "required": true,
            "schema": {
              "title": "Repertoire Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PracticeSessionOut"
                  },
                  "title": "Response List Practice Sessions Api Repertoire  Repertoire Id  Practice Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Practice Sessions",
        "tags": [
          "practice rooms"
        ]
      }
    },
...
    "/api/rooms": {
      "post": {
        "operationId": "create_room_api_rooms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Room",
        "tags": [
          "practice rooms"
        ]
      }
    },
    "/api/rooms/{room_id}": {
      "get": {
        "operationId": "get_room_api_rooms__room_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "title": "Room Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Room",
        "tags": [
          "practice rooms"
        ]
      }
    },
...
  }
}